Moved FakePlayerData to the capabilities package.
